requirements.txt
python-dotenv
azure-identity
openai
faiss-cpu
numpy
rich